[CHORE] Animation 파일로 분리(#524)
